Clean up, add comments
